move action base class to entity action lib folder
