+++ libs/entity-action/index.ts
@@ -1,2 +1,3 @@
+export * from './action';
 export * from './entity-action';
 export * from './entity-bulk-action';
